Specfn - Update to latest semi tested version (#18)
* Renamed variables, added imfp user choice, added internal theta calculation
Renamed write_photo_output to make the name more in line with actual effect.
Added user choice for IMFP (constant, layer dependent constant, energy dependent curve)
Renamed photo_imfp_const to better reflect it holding the values rather than just a constant
Added an array of internal theta values, since they change at the interface due to the change in E_kin.
Changed the calculation of the escape probability to the energy of the final band in the 3step model rather than the initial band+E_photon. The QE is thus now calculated with the escape probability of the final state rather than the escape probability of the initial state.

* Added od2od support for transmission coefficients

* resolved uncertainties regarding emission angle reference

* Added Electron Transmission Probabilities
These describe the probability for an electron in a bloch band to pass through the surface and across the potential step
The probabilities are calculated in CASTEP and supplied as a .tmprob_bin file.

* Added conversions in od2od for transmission probabilities

* Added support for supercell spectral function projections

* Corrected variable for MPI reading of specfn_bin

* Added some extra param printing

* Added cmd line option to supply max_gvec

* Added sane default for bind_energy smearing

* WIP - Reduced QE array dimensions to reduce memory requirements

* Revert "WIP - Reduced QE array dimensions to reduce memory requirements"

This reverts commit 9f9463ea4b7852c006d70f82091b339b0acc2224.

* Combined layer and bulk slab escape prob arrays

* Removed unused variables

* Fixed bug in combined emission probability implementation
The light intensity at the "bulk layer" was incorrectly used twice in the QE equation.

* WIP - Reduced dimension of 1 Step QE array
Needed to save memory so reducing the dimensionality of the big QE array. Need for testing as some more weighting has to be done for the correct calculation of the MTE

* WIP - 1 Step QE array dimension reduction

* WIP - Possibly correct approach

* WIP - Moved E_transverse contribution calculation into main QE loop

* Fixed deallocation error

* WIP - Reduced 3 Step QE array's dimensions

* Added comments for soon TODO

* Started implementation of new binding energy broadening

* Changed binding energy broadening for 1step
More work has to be done now, but the memory requirements should be much less overall

* Earlier dellocation of Etrans array should reduce memory requirement

* Reverted limiting binding_temp calculation

* Removed bug in summation process

* Moved g+k dependent gaussian value into the correct loop

* Added iteration reduction to 1step binding energy

* Untested implementation for 3step broadened binding energy

* removed compilation errors

* WIP - 3step binding energy with spectral function

* Binding Energy 3step Bug removed - not yet parallelised

* Removed extra printing statements and some old variables

* Checked parameters for MPI parallelisation and reordered photo params

* Flipped indices in transmission probability array

* Moved parallel data exchange flow for binding energy plots

* Added extra emission angle data to binding energy plot file

* Resolved missing variable conflicts and removed unused variables

* fprettify run

* Removed already commented out (redundant) and or unused functionality

* lines removed

* Fprettify on the codebase

* removed extra variables