Geometry cleanup (#19)
* Started cleanup of old geometry definition - untested, compiles succesfully

* Removed array index bug for I_layer

* Readded Atom label capitalisation

* removed some old variables and added minor printing

* added scalable printing from photo_dev branch

* Added some extra comments, removed commented out code
A lot of the removed code was used originally to define layers of atoms to be used. Now the layers are
defined by boxes of material with atoms in them and the optical properties are calculated for each of those boxes.
The QE values are still calculated for a every single atom.